CSP-2142 Rename all location-related fields to LocationName across codebase, database schema, stored procedures, and tests.

+++ src/SFA.DAS.Roatp.Jobs/ApiModels/Lookup/LocationAddress.cs
@@ -2,6 +2,6 @@
 
 public class LocationAddress
 {
-    public double? Latitude { get; set; }
-    public double? Longitude { get; set; }
+    public decimal? Latitude { get; set; }
+    public decimal? Longitude { get; set; }
 }
